move model impls to a file that is included by feature
